Add scheduling of federated batch jobs
Sibling jobs have to get lock from the origin cluster in order to attempt to
allocate nodes. If it gets the allocation then it lets the origin cluster know
and the origin cluster will set the siblings jobs, if any, into a REVOKED state
and purge the jobs. If the sibling job is the only sibling then it assumes the
lock and attempts to start the job to avoid extra communications. If nodes can't
be allocated then the job releases the lock for another cluster to try.