Alter configure check for libmunge to explicitly look for the .so file.
Configuring slurm with munge manually installed in /usr/local, with the
library in /usr/local/lib but an empty /usr/local/lib64 directory
will cause the munge plugins to look for libmunge.so in the wrong place.

The munge.spec file has historically provided libmunge.so as part of
munge-devel, which Slurm depends on already.