Remove local definition of PATH_MAX if macro was missing.
PATH_MAX is used in many other places within the source without these
definitions; while technically optional in POSIX1.2008 I have not
seen anywhere that does not set it, and if such a system exists
it'll be needed elsewhere and the build system should handle it.

Also, POSIX1.2008 states this must be at least 1024, not 256
as used here.