Fix regression from commits e3a7bdcc47e f9804256bac d72b13f220a
Reference bug 3366

If you are running on a Bluegene system we rely on the prolog to take us out of configuring
state.  These commits work good for system rebooting the nodes where the prolog is running,
but in the case of Bluegene this is the opposite desire :).   These commits on a Bluegene
pretty much make it so a batch job never gets launched.