Put clsuter_name in SLURM_WORKING_CLUSTER env
Previously it was just putting the IP and the srun would check to see if
slurm.conf's cluster name and the IP were the same to prevent setting
the working cluster unless if needed to. But since it was was checking
the cluster name and the IP it would always try to set the working
cluster rec even if they were the same.