Add --with-shared-libslurm to configure to determine if the system should be build linking dynamically or statically to libslurm.
This can result in much smaller binaries, but isn't the easiest to develop against as if the files in
src/common or src/*api change and the binaries linking against them are compiled directly
libslurm.so doesn't get recompiled automatically, you have to manually do it.  But in production
this could give large benefits as Slurm's footprint is now considerably smaller.

This used to not work on environments like AIX where plugins couldn't resolve variables used in
the parent program, but this appears to not be the case now we link to the "full" .so which exports
everything as the .o does.