Remove obsolete logic regarding job exit_code meaning.
There are (or at least used to be) places where an exit code of 0 was used
to indicate that the job ran successfully to completion. This patch fixes
the places where the exit_code was MAX'd out to 1 to indicate that job did
not ran successfully to completion. This MAX logic was introduced in commit
f305f099.

Bug 3444