Allow non-slurm users to submit federated jobs
The problem was that the fed_mgr sets the jobid before the jobid is
validated so that the jobid can be the same when submitting to all
siblings. _validate_job_desc() validates that only slurm_user or root
can specify jobids. Now in a federation, _validate_job_desc() doesn't
need to validate the jobid since specifying a specific jobid is disabled
in federations.