Send allocation's resp_host to fed siblings
When an interactive allocation request comes to a controller it fills in
the job's resp_host from the incoming addr. The controller then uses the
resp_host and the alloc_resp_port, sent from srun/salloc, to respond to
listening srun/salloc. In a federation, the origin cluster needs to pass
the initial resp_host from the origin cluster to the siblings. Otherwise
the siblings set the resp_host to the host of the origin cluster and the
sibling clusters won't be able to contact the listening srun/salloc.