Remove test15.17
Continuation of 0098c0c05479dde -- which removed the ability submit a
batch step within an existing allocation. Removing test15.17 since the
functionality was removed. This is undocumented behavior and says it was
for LSF which isn't supported. There is also the problem where if you
submit two batch steps in an exsiting allocation that the job will be
killed and the node drained because the slurmd will see duplicate
jobids.