Fix regression in 16.05.10 with respects to GrpTresMins on a QOS or Association.
This reverts commits 92d2c645843 and 37be42ec846.  This caused incorrect
behavior, original code was correct.  This also corrects documentation additions
in commit 4cfe6bde08fd.

This code caused the first clause never to be correct and if you were
over the limit you would get the third clause reporting a huge number available
where it should be a negative number.  The reality is the first clause should
had been triggered and handled correctly.