Fix controller/cmds talking to a pre-released DBD
A 17.02 controller,sacctmgr couldn't talk to a "master/17.11" DBD
because the 17.02 client was talking attempting to talk to the DBD with
the 17.02's MIN_PROTOCOL_VERSION -- which was 15.08 and is more than 2
version behind the master. The master's MIN_PROTOCOL_VERSION is 16.05,
so it couldn't unpack the messages.

The controller should always communicate at it's current protocol to the
DBD.

For federations, it's possible that a higher version controller could
talk to a lower version controller. So the cluster needs to talk to the
remote cluster using the remote cluster's protocol version -- which is
given back from the DBD.