Add NoDecay flag to QOS.
A QOS with this flag set will not have its GrpTRESMins,
GrpWall or UsageRaw decayed by the global PriorityDecayHalfLife or
PriorityUsageResetPeriod values.  Such a QOS will act as a time-limited
quota of resources.

This flag only affects the QOS and not the associations that run in the
QOS.  Those usage values will continue to be decayed.

The only way to reset this value is with

sacctmgr mod QOS $qos set rawusage=0

If you remove the flag usage will begin to be decayed as normal.