Do not set cpu frequency / governor on batch or extern steps.
This now matches the behavior documented in sbatch.

This resolves a problem where the maximum cpu frequency would be set
to the minimum available on the node by the batch step. This is due
to the batch step leaving cpu_freq_{min,max,gov} uninitialized to zero,
which is then translated to a request to set the frequency to the lowest
available in the node. This did not impact 16.05 or earlier, as a request
for a zero frequency was ignored by a quirk of _cpu_freq_freqspec_num.
This quirk was removed by commit f40e1c013d0 before 17.02.0-rc1.

Bug 3510.