Fix pam_slurm_adopt.
The stepd->protocol_version field is not set at this point,
if not handled here then the following call to stepd_get_uid
will fail as a protocol_version of zero is rejected.

This appeared to work in 16.05 and older as the stepd_get_uid()
API call used to ignore the invalid protocol_version and
return the correct result anyways, but that was fixed in c177ff954eb.

Bug .