Consider QOS flags Partition[Min|Max]Nodes when doing backfill.
Bug 3530

NOTE: there does appear to be other issues here, but we didn't fell comfortable
changing this many things in 17.02 for fear of breaking something.

This only fixes a bit of the issue as it appears node_scheduler.c has a fuller
test for these.  In 17.11 I plan to make this a function that will fill in
the min, req, max nodes and use it in both places in the code to prevent this
from happening again.