This is the meat of getting the consumed energy for the job to the database.
We use the last step in the allocation (works for everything but salloc) to
get the data to the job.  We update the job as steps finish and send a new
tres_alloc_str to the database which will update the job record with the info.

We could not use the job_comp message for anything but an salloc since there
is not guarantee it will get there after the last step does.