OK, this one is what we are all waiting for.
What this does is removes the join on the step table during rollups.  It turns
out this was a very costly join if the tables were fairly large.

We can now just grab the data directly from the job_table which already contains
the data in the tres_alloc column.