Fix slurmdbd_defs.c to not have half symbols go to libslurm.so and the other half go to libslurmdb.so.
Turns out in 17.11 if you are using --with-shared-libslurm sacct will link to
both libslurmfull.so and libslurmdb.so.  In the case of linking to the
accounting_storage/slurmdbd plugin it will callback to slurmdbd_defs_init
which would be in libslurmdb.so but the call to acct_storage_p_get_connection
would call slurm_open_slurmdbd_conn which is in libslurmfull.so and it would
xassert on slurmdbd_defs_inited which was set in libslurmdb.so but not in
libslurmfull.so.

So the moral of the story is don't export half a file in one lib and the other
half in a different lib.

Perhaps we should only have 1 lib all together, but that isn't the way it is
done today.  This fixes the issue and has the entire file exported to
libslurmfull.so so we should be good.  But just a note for the future.

This was an unexpected regression caused by commit 5a5347c75f1371.