Fix packing federation's sib_msg buffers
We were changing the buffers' processed offset to the beginning of the
body part of the offset to pack just the body part of the buffer but
once this happened we lost how much actual data was in the buffer. So we
were packing the whole buffer which could be larger than what was
actually being used in the buffer.