Fix conversion table when there are no clusters
If there were no clusters in the system then the dbd would insert a row
in the database and then on the next time the dbd would start (with no
clusters) it would attempt to insert the row again and fatal. Now it
always makes a query against the database and if there are no clusters
in the system, then it inserts the current db version into the table.