Properly remove clusters when removed from fed
In 6eec802, the cluster's recv connection is now being destroyed when
the cluster is being destroyed. The problem that showed itself was that
when a remote cluster is removed from the federation, the controller
calls slurmdb_destroy_federation_rec() which destroys the cluster's in
the list. Both the persistent recv thread and the cluster's recv are
pointing to the same thing so when the controller removed the recv
persistent connection the recv thread was pointing to bad memory.