Set revoked job's start time to end time
Mimicking how cancelled jobs are. The database will show that the job
start_time is 0 but in the controller the the start time will be the
same as the end time. sacct will set the start time to the end time if
there is an end time and the start time is 0.