Filter out duplicate federated jobs
When running sacct from a federated client, the db returns jobs for each
cluster with duplicate jobs removed on each cluster. A federated job
could have ran on a different cluster when the before the jobid's
rolled. This patch filters out past old federated jobs and leaves the
newest ones.