Prevent scontrol crash when operating on array and no-array jobs at once.
When requesting an operation on jobs, where the operation permits to specify
more than one job in the same request, and a job array appears before a
regular job (no-array job) in the list of jobs to operate with, the
job_array_resp_msg_t pointer was not properly NULL'ed and thus incorrectly
accessed when processing the no-array job. This fix prevents the crash from
happening in the following scontrol operations:

uhold, hold, suspend, requeue, requeuehold, update, release

when the same request has <array_jobid>,<non-array_jobid> in this order in
the job list to process.

Bug 3759