Open fed persist conns at time of join federation
Previously it was only opening the connections at startup and then if
had something to send to a sibling that it didn't have a connection to.
It's better to have them establish the connection when the cluster is
added to the federation so that it can detect if the cluster is down
later and the clusters can sync jobs when joing the federation.