Improve slurmd startup on large systems (> 10000 nodes)
What this does is populate the node_hash_table as nodes are being read in
instead of after the node_record_table_ptr has been fully populated.

This speeds up a start of a slurmd with a system of 10000 nodes from
> 1 minute to less than a second.

In 17.11 we will remove the linear xstrcmp check as it should no longer be
needed.

Bug 3885