Update federation syncing.
Handle syncing between siblings (not the origin). Job could have been
canclled or started while the origin was down. The sibling should see
this and remove it's copy if the job is running or cancelled on the
remote cluster.

Handle case where job was cancelled or finished while the origin was
down and the siblings were up.