CRAY - Alter algorithm to come up with the SLURM_ID_HASH.
The SLURM_ID_HASH used for Cray systems has changed to fully use the
entire 64 bits of the hash.  Previously the stepid was multiplied by
10,000,000,000 to make it easy to read both the jobid as well as the
stepid in the hash separated by at least a couple of zeros, but this
lead to overflow on the hash with steps like the batch and extern step
where they used all 32 bits to represent the step.  While the new method
ruins the easy readability it fixes the more important overflow issue.
This most likely will go unnoticed by most, just a note of the change.