Let remote fed jobs stay in queue till minjobage
Previously remote jobs would be removed from the job_list as quickly as
possible to prevent collisions with requeued jobs and to clear up the
jobs and the orign job would stay around till minjobage on the origin.
But the origin job didn't have the details from the job that ran on a
remote cluster.

Now just don't show revoked jobs. The origin tracking job will remain as
revoked and not shown and the remote job will hang around for display
till minjobage. scontrol show jobs will show the job from the cluster
that ran the job. The job is requeuable as long as the origin job is
still in the origin cluster's job_list.