Fix regression in commit e5c05549657124 that would put the stepd pid into the memory cgroup instead of the task's pid.
Beforehand this would put the result of getpid() into the cgroup.  Before
e5c05549657124 this was done in the child of the fork which would get you
the task's pid, but moving it to run in the parent broke this logic.

What this patch does is adds pid to the input parameters of
task_g_pre_launch_priv making it so we could use the correct pid.