Fix issue with whole gres not being printed out with Slurm tools.
This is a regression from commit fec995e0cd087.

It turns out using tok here was erroneous for situations where the gres had
a type and name and potentially a count (i.e. network:gigabit:1)

_get_gres_req_cnt() would alter the incoming char *config which is what tok
was.  So when we print it back to the requested string it would only have
what was there to the first ':'.  As we didn't need to \0 out the first char
as we skip over it anyway I just kept track of what the replaced \0 was for
the number portion and put it back when we are done copying it.

Related to bug 3521