Fix uninitialized value
reported when compiling with optimizations (-O2). field_id may be
uninitialized or the the value from the previous iteration in the while
loop. The only possible values of dataset_loc->type are:

typedef enum {
	PROFILE_FIELD_NOT_SET,
	PROFILE_FIELD_UINT64,
	PROFILE_FIELD_DOUBLE
} acct_gather_profile_field_type_t;

and the while loop condition ensures that PROFILE_FIELD_NOT_SET is not
handled. So instead of handling PROFILE_FIELD_NOT_SET directly, just
catching everything else with the "default" case statement and
continuing the loop.