Add SLURM_DROP_PRIV to flags in slurm_msg_t to drop priviledges when testing.
Handling is enabled only in development builds. When set in the message
header, a thread-local variable in slurmctld is set to force the auth
functions (validate_slurm_user() et al) to treat the user as unpriviledge.

Designed to be used along with the testsuite at a later point in time, such
that tests can validate user access restrictions without needing access to
alternate accounts (since the testsuite is usually run as SlurmUser or at
least an Admin to validate a lot of administrative components).

Bug 4080.