Rearrange where drop_priv and drop_priv_flag live.
The drop_priv symbol needs to be accessible from common_as.c,
which is compiled in with the accounting_storage/mysql plugin.

Since that plugin can run in slurmdbd, or directly from slurmctld and
the user commands, the only way to get drop_priv there consistently
is to move the declaration to some common ground behind all of them.

Rather than expose this throughout, just explicitly import it with
'extern' in the appropriate locations.