Add slurm_thread_create and slurm_thread_create_detached macros.
Set to fatal on pthread_create failure, as this is the LCD of all
existing behaviors, and as such if something were stopping thread
creation from succeeding you could expect some code path to trip
that eventually. Rather than gamble on when/if that is triggered,
just immediately die.

Bug 4163.