slurm.spec: move a copy to contribs/slurm.spec-legacy
Preserve an older copy of slurm.spec. The shipped version
will be overhauled, and targeted at RHEL 7+ / SuSE 12+ only
and assume systemd is being used.

This copy is deprecated, not maintained, and will be removed
in a future release.

While relocating, update the header block with the package name,
correct the license description, and set the release.

Important note: do not be tempted to rename this to slurm-legacy.spec.
Doing so will confuse rpmbuild, as it attempts to build everything in
the archive file with a .spec extension on it and this will immediately
conflict with the shipped version of slurm.spec.

Start of work for bug 3396.