Add empty hashes in perl api for hidden nodes
The contoller will return node records with a NULL name for nodes that
are hidden. This is so that you can map a partition_info's nodes --
using it's node_inx[] -- to a node record in the returned node_array
from slurm_load_node().

Previously the perl api would leave an undefined object in the the
node_array if the hidden nodes were found before a real node in the
node_array and any hidden nodes at the end of the array from the
controller wouldn't be counted for in the perl node_array. This patch
adds empty hashes for hidden nodes and preserves the record_count and
node_array from the slurmctld.

Bug 4250