Fix another uninitialized warning.
Move test above the accounting limit check. Change the test
to verify first_job_ptr instead of pack_job_id to avoid warning
about a possible dereference of an uninitialized variable lower down.
(Note that both pack_job_id and first_job_ptr are set at the same time,
so this test is functionally equivalent but able to convince the compiler
that things will be okay.)