Update spank expect tests 7.11 and 38.6
The problem was the tests set the same variable, opt_arg, for for
options, --test_suite and --test_suite_reg. But with the addition of
de24ce72dc16 the variable was being set with the different numbers out
of order because the variable was being set from the environment. So the
fix is to have the options set their own variable.

Bug 4176