Fix issue when resetting the partition pointers on nodes.
Issue could be triggered when updating a partition node(s) with node(s)
that were already in the partition, incorrectly increasing the
node_record->part_cnt (number of associated partitions) and thus
incorrectly extending the array of pointers to partitions associated
with this node, leading to an array with repeated associated
partitions pointers.

Bug 4318.