slurm.spec - do not generate and install an ld.so.conf.d file.
%{_libdir} is /usr/lib64. Adding this can cause ordering
problems with the library search path, and is redundant.

And %{_libdir}/slurm is not needed by Slurm itself; it relies on
rpath to resolve the install location instead.

So, this is only useful on a non-standard install, in which case
it's best left to the site to decide how to handle this instead.

Bug 4344.