Fix test to handle output if allocation is delayed.
This primarily happens regularly if

PrologFlags=contain

Where the output has a nodename (and usually a number in it)

Normal output is

salloc: Granted job allocation 31159
1

but if the allocation is delayed you get output like

salloc: Granted job allocation 31161
salloc: Waiting for resource configuration
salloc: Nodes snowflake4 are ready for job
1
salloc: Relinquishing job allocation 31161

Where the test would key on the number in snowflake4 and fail.

This change makes it so the number has to be the first thing seen.