Show federated jobs over local cluster's copy
If a pending federated job exists on clusters 2 and 3 and squeue is run
from cluster 1 then the active siblings can come and go depending on
which cluster returns the job info first and depending if that cluster
is the origin cluster or not. The origin cluster only knows where the
active siblings are.