Fix for commit 9ed24025a6 to fix issue where job ordering causes wrong unused wall.
The "order by" part of the original sql query was wrong.
I had a situation where I had 2 jobs in an updated reservation,
but instead of adding their times together, they were overwriting
their times. This happened because of the ordering of the rows, where
the reservation got split up because of the "order by job.time_start".

That was simply wrong.

Bug 4406