Fix sbatch --wait to stop after job is gone
slurm_load_job() prior to 17.11 returns the error code in errno and not
in rc.  With the addition of 47175901f if a job is removed from memory
before sbatch checks for the job again, sbatch could get in an loop
checking for the job. This only happens if you have a very small
MinJobAge (<10) -- which is not recommended.