Fix to properly remove extern steps from the starting_steps list.
Since NO_VAL = SLURM_BATCH_SCRIPT, the else statement would only compare
the job_id and not the step_id, thus when a batch step was removed all
the steps from that job would be removed too. Then when attempting to
remove the extern step in the next iteration, it was already removed
and we were incorrectly erroring out.

Bug 4458.