Replace slurm_playbook.yaml.in with a version that auto-detects the Slurm version.
Due to the way Cray builds Slurm, the prefix and bindir paths include the
Slurm version (/opt/slurm/<version>). This means every time we update to a
new Slurm version we must update the Slurm ansible playbook. It also means
that the slurm_playbook.yaml file must be built with Slurm to be used
(it can't simply be copied directly).

The attached patch updates the playbook to determine the version of Slurm
to use from the module file, and hardcodes the sysconfdir setting we give
in our Slurm installation guide. If a customer uses different paths, they
can update the playbook to meet their needs.

Bug 4360.