Update jobcomp records on duplicate inserts
In the federation case, the origin job is completed in the database when
a sibling job starts the job. The complete message is then sent again to
the database when the job is completed on the sibling cluster but
it is updated with the sibling job's exit code.

The jobcomp plugin didn't handle the multiple updates to the record.
This change allows the existing record to be updated.

Bug 4493