slurm.spec - move libpmi libraries to a separate package.
The slurm and pmix RPMs both attempt to install a libpmi. While Slurm's
version should be preferred, the best way to handle this is to have both
pmix and slurm packages split this off into a separate RPM that can
conflict with each other, giving the admin a way to resolve the conflict
manually.

Bug 4511.