Improve logic when summarizing job arrays mail notifications.
When --mail-type option isn't requested with ARRAY_TASKS, we need somehow
to summarize the different states each task finished in the array. We've
added a new ARRAY_TASK_REQUEUED flag to the array_flags to indicate that
at least one task was requeued. Also the logic now detects if at least
one task failed and/or if otherwise all finished successfully.

The patch also removes the RunTime from the the e-mail when summarizing
whole array, since it doesn't make sense to specify just the RunTime
of one of the tasks for this case.

Patch also fixes when ARRAY_TASKS is specified, previously the mail
notification for the master job task record included a range of
ExitCodes for all the tasks. Since this option is not for summarizing,
the patch makes it so only the range is shown when the option isn't
specified.

Bug 4539.