Before, srun allowed execution when checking a command existing in the current directory but not in the --chdir specified cwd.
Now srun looks for the command in specified cwd and differentiates all
the possible cases where a command starts with ./ (relative), with /
(absolute), and with a different char indicating a command that is in
the path.

Fix also for a PATH variable starting by '.', now it is relative to
specified cwd.

bug 4392