Prevent multiple calls to pthread_atfork() after re-init of config regex.
keyvalue_initialized is reset on 'scontrol reconfigure' and other
cases, which can lead to additional atfork handlers being registered.

These can eventually lead to a segfault if an excessive number of
handlers have been re-registered. Set a separate boolean to protect
against this. Clear that boolean as part of the atfork handler.

Bug 4628.