Add fatal_abort() function to log a last fatal message then abort.
fatal() calls exit(1) which precludes getting a backtrace.
That's fine on configuration issues and other types of problem,
but for hitting "impossible" edge cases getting a core dump may
be the only way to isolate the issue.

Adding to 17.11 so we can easily provide diagnostic patches without
needing users to back-port this implementation. Further use will
come in 18.08.

Bug 4599.