Preserve and fix node features on reconfig or restart
This patch fixes the situation that makes features unrecognized where a node
features plugin is active and features are defined to nodes in slurm.conf.

It also preserves KNL node features when slurmctld daemons are reconfigured
including active and available modes.

Features not belonging to node features plugin are reset to what is in
slurm.conf when restarting or reconfiguring.

Bug 4734